+++ lnmp/lnmp.sh
@@ -34,8 +34,13 @@ read -p " please select whitch item u want to install:  " itemIdx
 case $itemIdx in
     1)
         echo > "$MYSQL_INSTALL_LOG"
-        echo -e $BR
-        read -p "please input the mysql install dir: " MYSQL_INSTALL_PATH
+
+        if [ $INSTALL_MODE -eq ${CUSTOM_MODE} ];then
+            echo -e $BR
+            read -p "please input the mysql install dir: " MYSQL_INSTALL_PATH
+        else
+            MYSQL_INSTALL_PATH="mysql"
+        fi
         
         MYSQL_INSTALL_PATH="$INSTALL_PATH"/"$MYSQL_INSTALL_PATH"
 
@@ -58,8 +63,36 @@ case $itemIdx in
     ;;
     2)
         echo > "$NGINX_INSTALL_LOG"
-        echo -e $BR
-        read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
+
+        #chose the install item
+        echo -e $TOP_MARK
+        echo " install nginx   --- 1 "
+        echo " install tengine --- 2 "
+        echo -e $BOTTOM_MARK
+        read -p "  please select whitch package u want to install: " idxItem
+        
+        if [ "$idxItem" -eq 1 ];then
+            NGINX_TYPE=1
+        elif [ "$idxItem" -eq 2 ];then
+            NGINX_TYPE=2
+        else
+            echo -e $TOP_MARK
+            echo " please input a right num "
+            echo -e $BOTTOM_MARK
+            exit 1
+        fi
+
+        if [ ${INSTALL_MODE} -eq ${CUSTOM_MODE} ];then
+            echo -e $BR
+            read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
+        else
+            if [ ${NGINX_TYPE} -eq 1 ];then
+                NGINX_INSTALL_PATH="nginx"
+            elif [ ${NGINX_TYPE} -eq 2 ];then
+                NGINX_INSTALL_PATH="tengine"
+            fi
+        fi
+        export NGINX_TYPE
         
         NGINX_INSTALL_PATH="$INSTALL_PATH"/"$NGINX_INSTALL_PATH"
 
@@ -81,9 +114,39 @@ case $itemIdx in
     ;;
     3)
         echo > "$PHP_INSTALL_LOG"
-        echo -e $BR
-        read -p "please input the php install dir: " PHP_INSTALL_PATH
-        read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
+        
+
+        if [ ${INSTALL_MODE} -eq ${CUSTOM_MODE} ];then
+            echo -e $BR
+            read -p "please input the php install dir: " PHP_INSTALL_PATH
+            read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
+        else
+            #chose the install item
+            echo -e $TOP_MARK
+            echo " install nginx   --- 1 "
+            echo " install tengine --- 2 "
+            echo -e $BOTTOM_MARK
+            read -p "  please select whitch nginx package u had already installed: " idxItem
+        
+            if [ "$idxItem" -eq 1 ];then
+                NGINX_TYPE=1
+            elif [ "$idxItem" -eq 2 ];then
+                NGINX_TYPE=2
+            else
+                echo -e $TOP_MARK
+                echo " please input a right num "
+                echo -e $BOTTOM_MARK
+                exit 1
+            fi
+            export NGINX_TYPE
+            
+            if [ ${NGINX_TYPE} -eq 1 ];then
+                NGINX_INSTALL_PATH="nginx"
+            elif [ ${NGINX_TYPE} -eq 2 ];then
+                NGINX_INSTALL_PATH="tengine"
+            fi
+            PHP_INSTALL_PATH="php"
+        fi
         
         PHP_INSTALL_PATH="$INSTALL_PATH"/"$PHP_INSTALL_PATH"
         NGINX_INSTALL_PATH="$INSTALL_PATH"/"$NGINX_INSTALL_PATH"
@@ -111,10 +174,40 @@ case $itemIdx in
         echo > "$MYSQL_INSTALL_LOG"
         echo > "$NGINX_INSTALL_LOG"
         echo > "$PHP_INSTALL_LOG"
-        echo -e $BR
-        read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
-        read -p "please input the mysql install dir: " MYSQL_INSTALL_PATH
-        read -p "please input the php install dir: " PHP_INSTALL_PATH
+        
+        #chose the install item
+        echo -e $TOP_MARK
+        echo " install nginx   --- 1 "
+        echo " install tengine --- 2 "
+        echo -e $BOTTOM_MARK
+        read -p "  please select whitch nginx package u want to install: " idxItem
+        
+        if [ "$idxItem" -eq 1 ];then
+            NGINX_TYPE=1
+        elif [ "$idxItem" -eq 2 ];then
+            NGINX_TYPE=2
+        else
+            echo -e $TOP_MARK
+            echo " please input a right num "
+            echo -e $BOTTOM_MARK
+            exit 1
+        fi
+
+        if [ ${INSTALL_MODE} -eq ${CUSTOM_MODE} ];then
+            echo -e $BR
+            read -p "please input the mysql install dir: " MYSQL_INSTALL_PATH
+            read -p "please input the php install dir: " PHP_INSTALL_PATH
+            read -p "please input the nginx install dir: " NGINX_INSTALL_PATH
+        else
+            if [ ${NGINX_TYPE} -eq 1 ];then
+                NGINX_INSTALL_PATH="nginx"
+            elif [ ${NGINX_TYPE} -eq 2 ];then
+                NGINX_INSTALL_PATH="tengine"
+            fi
+            MYSQL_INSTALL_PATH="mysql"
+            PHP_INSTALL_PATH="php"
+        fi
+        export NGINX_TYPE
         
         NGINX_INSTALL_PATH="$INSTALL_PATH"/"$NGINX_INSTALL_PATH"
         MYSQL_INSTALL_PATH="$INSTALL_PATH"/"$MYSQL_INSTALL_PATH"
